First commit, for real this time!
I've tried getting this project going a few times on gitHub, but being a newb and not learning how to use git properly before committing my code, it hasn't worked out so well. This is an early version of Pong that isn't quite finished yet. The game never ends at this point, and score doesn't effect anything. The AI is too smart for its own good and there are no menus. It works, though.